+++ tests/test_configuration.py
@@ -8,7 +8,7 @@ from waveform_editor.tendencies.linear import LinearTendency
 from waveform_editor.tendencies.periodic.sine_wave import SineWaveTendency
 from waveform_editor.tendencies.smooth import SmoothTendency
 from waveform_editor.waveform import Waveform
-from waveform_editor.yaml_parser import YamlParser
+from waveform_editor.yaml.yaml_parser import YamlParser
 
 
 @pytest.fixture
