+++ tests/test_exporter.py
@@ -3,7 +3,7 @@ import numpy as np
 import pytest
 
 from waveform_editor.configuration import WaveformConfiguration
-from waveform_editor.exporter import ConfigurationExporter
+from waveform_editor.export.exporter import ConfigurationExporter
 
 
 @pytest.fixture
