+++ waveform_editor/gui/dict_editor.py
