+++ waveform_editor/gui/shape_editor/__init__.py
@@ -1,0 +1,3 @@
+from waveform_editor.gui.shape_editor.shape_editor import ShapeEditor
+
+__all__ = ["ShapeEditor"]
