+++ waveform_editor/gui/shape_editor/coil_currents.py
