+++ waveform_editor/gui/shape_editor/nice_plotter.py
@@ -11,9 +11,9 @@ from imas.ids_toplevel import IDSToplevel
 from panel.viewable import Viewer
 
 from waveform_editor.settings import NiceSettings, settings
-from waveform_editor.shape_editor.nice_integration import NiceIntegration
-from waveform_editor.shape_editor.plasma_properties import PlasmaProperties
-from waveform_editor.shape_editor.plasma_shape import PlasmaShape
+from waveform_editor.gui.shape_editor.nice_integration import NiceIntegration
+from waveform_editor.gui.shape_editor.plasma_properties import PlasmaProperties
+from waveform_editor.gui.shape_editor.plasma_shape import PlasmaShape
 
 matplotlib.use("Agg")
 logger = logging.getLogger(__name__)
