+++ waveform_editor/gui/shape_editor/xml_param/direct_param.xml
