+++ waveform_editor/gui/shape_editor/xml_param/inverse_param.xml
